requirements.txt
discord.py>=2.3
Flask>=2.2